+++ crates/metadata/src/format.rs
@@ -1,0 +1,21 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Format {
+    #[serde(default = "default_indent_width")]
+    pub indent_width: usize,
+}
+
+const DEFAULT_INDENT_WIDTH: usize = 4;
+
+impl Default for Format {
+    fn default() -> Self {
+        Self {
+            indent_width: DEFAULT_INDENT_WIDTH,
+        }
+    }
+}
+
+fn default_indent_width() -> usize {
+    DEFAULT_INDENT_WIDTH
+}