+++ crates/metadata/src/project.rs
@@ -1,0 +1,13 @@
+use semver::Version;
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Project {
+    pub name: String,
+    pub version: Version,
+    #[serde(default)]
+    pub authors: Vec<String>,
+    pub description: Option<String>,
+    pub license: Option<String>,
+    pub repository: Option<String>,
+}