+++ crates/tests/src/lib.rs
@@ -32,7 +32,7 @@ mod analyzer {
 
         let ret = Parser::parse(&input, &file).unwrap();
         let prj = vec![&metadata.project.name];
-        let analyzer = Analyzer::new(&prj);
+        let analyzer = Analyzer::new(&prj, &metadata);
         let errors = analyzer.analyze_pass1(&input, &file, &ret.veryl);
         assert!(errors.is_empty());
 
@@ -103,7 +103,7 @@ mod emitter {
                 if path.src.starts_with(&cache_dir) {
                     let input = fs::read_to_string(&path.src).unwrap();
                     let ret = Parser::parse(&input, &path.src).unwrap();
-                    let analyzer = Analyzer::new(&path.prj);
+                    let analyzer = Analyzer::new(&path.prj, &metadata);
                     let _ = analyzer.analyze_pass1(&input, &path.src, &ret.veryl);
                 }
             }
@@ -114,7 +114,7 @@ mod emitter {
 
         let ret = Parser::parse(&input, &file).unwrap();
         let prj = vec![&metadata.project.name];
-        let analyzer = Analyzer::new(&prj);
+        let analyzer = Analyzer::new(&prj, &metadata);
         let _ = analyzer.analyze_pass1(&input, &file, &ret.veryl);
         let _ = analyzer.analyze_pass2(&input, &file, &ret.veryl);
         let mut emitter = Emitter::new(&metadata);