+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/veryl/compare/v0.3.4...Unreleased) - ReleaseDate
 
+* [Added] lint for naming convention [#116](https://github.com/dalance/veryl/issues/116)
+
 ## [v0.3.4](https://github.com/dalance/veryl/compare/v0.3.3...v0.3.4) - 2023-02-08
 
 ## [v0.3.3](https://github.com/dalance/veryl/compare/v0.3.2...v0.3.3) - 2023-02-08